backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.1
pydantic-settings>=2.1.0
pandas>=2.1.3
openpyxl>=3.1.2
python-magic>=0.4.27
requests>=2.31.0
python-jose[cryptography]>=3.3.0
google-generativeai>=0.3.0
